+++ pom.xml
@@ -171,6 +171,7 @@
 				<configuration combine.self="override">
 					<excludes>
 						<exclude>**/com/coderace/sql/SQLController.class</exclude>
+						<exclude>**/com/coderace/model/dto/*</exclude>
 					</excludes>
 				</configuration>
 				<executions>
