+++ src/main/java/com/coderace/repository/DeliveryRepository.java
@@ -1,6 +1,6 @@
 package com.coderace.repository;
 
-import com.coderace.delivery.Delivery;
+import com.coderace.model.entities.Delivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
